Remove UELCHandler
This object was simply a collection of utility methods. The database override is not necessary.
