+++ uelc/main/helper_functions.py
@@ -1,21 +1,22 @@
-import json
-import hmac
-import hashlib
-import time
 from datetime import datetime
+import hashlib
+import hmac
+import json
 from random import randint
+import time
+
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
+from django.core.cache import cache
 from django.core.exceptions import ObjectDoesNotExist
 from django.http import HttpResponseRedirect, HttpResponse
 from django.shortcuts import get_object_or_404
-
 from pagetree.generic.views import generic_instructor_page, generic_edit_page
 from pagetree.models import Hierarchy, Section
 
 from gate_block.models import GateSubmission
-from uelc.main.models import CaseMap, Case
+from uelc.main.models import CaseMap, Case, CaseAnswer
 
 
 def get_cases(request):
@@ -184,3 +185,127 @@ def get_user_last_location(user, hierarchy):
         ul.delete()
 
     return section
+
+
+def get_vals_from_casemap(casemap_value):
+    vals = [int(i) for i in casemap_value if int(i) > 0]
+    return vals
+
+
+def get_part_by_section(section):
+    key = 'uelc.{}.get_part_by_section'.format(section.pk)
+    v = cache.get(key)
+    if v is not None:
+        return v
+
+    modules = section.get_root().get_children()
+    sec_module = section.get_module()
+    part = 0
+    for idx, module in enumerate(modules):
+        if module == sec_module:
+            part = idx
+
+    if part == 0:
+        v = 1
+    else:
+        v = float(2) + (part * .1)
+
+    cache.set(key, v)
+    return v
+
+
+def get_partchoice_by_usermap(usermap):
+    vals = get_vals_from_casemap(usermap.value)
+    part = 1
+    if len(vals) >= 2:
+        part = float(2) + (vals[1] * .1)
+    if len(vals) >= 4:
+        part = part + float((vals[3] * .01))
+    return part
+
+
+def get_p1c1(casemap_value):
+    return get_vals_from_casemap(casemap_value)[1]
+
+
+def can_show(request, section, casemap_value):
+    cmvl = list(casemap_value)
+    tree = section.get_tree()
+    section_index = [sec for sec in
+                     range(len(tree))
+                     if tree[sec] == section][0] + 1
+    try:
+        content_value = [int(i) for i in
+                         reversed(cmvl[0:section_index])
+                         if int(i) > 0][0]
+    except IndexError:
+        content_value = 0
+
+    return content_value
+
+
+def can_show_gateblock(gate_section, part_usermap,
+                       part_section=None):
+    if part_section is None:
+        part_section = get_part_by_section(gate_section)
+
+    can_show = False
+    if part_section == 1 or part_section == round(part_usermap, 1):
+        can_show = True
+    return can_show
+
+
+def p1pre(casemap_value):
+    p1pre = 0
+    vals = get_vals_from_casemap(casemap_value)
+    if len(vals) > 1:
+        p1pre = vals[0]
+    return p1pre
+
+
+def is_curveball(current_section, pageblocks=None):
+    block = None
+    if pageblocks is None and current_section is not None:
+        pageblocks = current_section.pageblock_set.all()
+
+    if pageblocks:
+        for pb in pageblocks:
+            block = pb.block()
+            if (hasattr(block, 'display_name') and
+                    block.display_name == "Curveball Block"):
+                return (True, block)
+
+    return (False, block)
+
+
+def is_decision_block(current_section, user, pageblocks=None):
+    if pageblocks is None:
+        pageblocks = current_section.pageblock_set.all()
+    for pb in pageblocks:
+        block = pb.block()
+        if (hasattr(block, 'display_name') and
+                block.display_name == "Decision Block"):
+            ss = block.submission_set.filter(user=user).last()
+            ca = None
+            if ss:
+                response = ss.response_set.filter(
+                    submission_id=ss.id).last()
+                ca = CaseAnswer.objects.get(answer=response.answer())
+            return (True, block, ca)
+    return (False, None, None)
+
+
+def is_next_curveball(section):
+    key = 'uelc.{}.is_next_curveball'.format(section.pk)
+    v = cache.get(key)
+    if v is not None:
+        return v
+
+    nxt = section.get_next()
+    is_cb = is_curveball(nxt)
+    if is_cb[0]:
+        cache.set(key, is_cb)
+        return is_cb
+    else:
+        cache.set(key, False)
+        return False
